sum of the 3 number  should be match with the last digit of number
This is the programatic solution for a whatsapp brain test question!!!!
i was just thought to implement it out in programm to give the solution.
here is the problem and below is the solution :)

problem:

last digt of 3 same number should is the sum of these 3 numbers
i call this brw-circle algorithem as the problem was showen to me as
first digit in blue circle
second as red circle and 3rd as white. and result should be white 3
digit color

i tried this for 5 digit numbers
solution is NULL :(

+#*
+#*
+#*
---
***

sum of the 3 number ("+#*" is one number)  should be match with the last
digit of number